+++ CHANGELOG.md
@@ -5,4 +5,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Configuration schema for address and phone fields.
+- GraphQL queries to retrieve those configurations (all countries or a single one).